move wavePropagation1d to own folder
