+++ src/SConscript
@@ -9,7 +9,7 @@ Import('env')
 # gather sources
 l_sources = [ 'solvers/roe/Roe.cpp',
               'solvers/fWave/FWave.cpp',
-              'patches/WavePropagation1d.cpp',
+              'patches/wavePropagation1d/WavePropagation1d.cpp',
               'setups/shockShock1d/ShockShock1d.cpp',
               'setups/damBreak1d/DamBreak1d.cpp',
               'setups/rareRare1d/RareRare1d.cpp',
@@ -28,7 +28,7 @@ env.standalone = env.Object( "main.cpp" )
 l_tests = [ 'tests.cpp',
             'solvers/roe/Roe.test.cpp',
             'solvers/fWave/FWave.test.cpp',
-            'patches/WavePropagation1d.test.cpp',
+            'patches/wavePropagation1d/WavePropagation1d.test.cpp',
             'setups/shockShock1d/ShockShock1d.test.cpp',
             'setups/damBreak1d/DamBreak1d.test.cpp',
             'setups/rareRare1d/RareRare1d.test.cpp',
